requirements.txt
langchain
langgraph
langchain-gigachat
psycopg2-binary
python-dotenv
salute_speech
SQLAlchemy
maxapi[webhook]
fastapi
uvicorn
jinja2
ymaps